requirements1.txt
absl-py==0.15.0
alembic==1.2.1
aniso8601==7.0.0
anyio==3.7.1
argh==0.26.2
astor==0.8.0
astunparse==1.6.3
audioread==2.1.8
Authlib==0.14.3
blinker==1.4
blis==0.4.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
charset-normalizer==3.3.2
clang==5.0
click==8.1.7
cryptography==3.1
cymem==2.0.2
decorator==4.4.0
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#sha256=65cea113d7509d5ab515583bbcbc9a31b4157bebbb73133d4110d2945476eaa4
exceptiongroup==1.2.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.2
h11==0.14.0
h2==4.1.0
h5py==3.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
ibm-cloud-sdk-core==1.0.0
ibm-watson==4.0.1
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.1
jiwer==3.0.3
joblib==0.13.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==18.1.1
librosa==0.7.0
llvmlite==0.29.0
Mako==1.1.0
Markdown==3.1.1
markdown-it-py==2.2.0
MarkupSafe==1.1.1
marshmallow==3.0.0b16
marshmallow-sqlalchemy==0.19.0
mdurl==0.1.2
more-itertools==7.2.0
murmurhash==1.0.2
mysql-connector-python==8.0.17
numba==0.45.1
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
pathtools==0.1.2
pkg_resources==0.0.0
plac==1.1.3
polling2==0.5.0
preshed==3.0.2
protobuf==3.19.6
pveagle==0.2.0
pvleopard==2.0.2
pyannote.core==5.0.0
pyannote.database==5.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.19
pydub==0.23.1
Pygments==2.17.2
PyJWT==1.7.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-speech-features==0.6
pytz==2019.2
PyYAML==5.1.2
rapidfuzz==3.4.0
regex==2023.12.25
requests==2.22.0
requests-oauthlib==1.3.1
resampy==0.2.2
rich==13.7.1
rsa==4.9
scikit-learn==0.21.3
scipy==1.5.4
shellingham==1.5.4
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
SoundFile==0.10.2
spacy==2.2.2
speechmatics-python==1.14.5
SQLAlchemy==1.3.8
srsly==0.2.0
StringGenerator==0.3.4
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
thinc==7.3.1
toml==0.10.2
tqdm==4.37.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
typing_extensions==4.7.1
urllib3==1.25.6
validate-email==1.3
wasabi==0.3.0
watchdog==0.9.0
websocket-client==0.48.0
websockets==11.0.3
Werkzeug==0.15.6
wrapt==1.12.1
zipp==0.6.0
